+++ src/datasources/__tests__/fetch-data-source-v2.test.ts
@@ -23,12 +23,12 @@ function makeDataBundle() {
     translations: {
       v: 1,
       data: {
-        headsigns: {},
-        stop_headsigns: {},
-        stop_names: {},
-        route_names: {},
         agency_names: {},
-        agency_short_names: {},
+        route_long_names: {},
+        route_short_names: {},
+        stop_names: {},
+        trip_headsigns: {},
+        stop_headsigns: {},
       },
     },
     lookup: { v: 2, data: {} },